+++ apps/next-client/src/app/[locale]/stripe/checkout/fail/page.tsx
